displaying element on the browser
